[Refactor/#135] TermsScreen을 screens.terms 패키지로 이동
